Sammelanteilsschein added. It is based on the template.

+++ Model/Umwandlungsfaktor.cs
@@ -5,6 +5,6 @@ namespace Anteilsscheine.Model
     public class Umwandlungsfaktor
     {
         public int Year { get; set; }
-        public float Factor { get; set; }
+        public decimal Factor { get; set; }
     }
 }
